+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - fixed regression in list that broke mwc-select in IE or shady dom.
 - Tabs no longer focus on initialization
 - mwc-list-item ripple color will now change based off of --mdc-ripple-color on initialization
+- Fix issue where textfield would throw an error when fed a non-string value
 
 ### Added